+++ src/main/java/no/vegvesen/nvdbapi/client/model/roadnet/route/RouteOnRoadNet.java
@@ -3,15 +3,28 @@ package no.vegvesen.nvdbapi.client.model.roadnet.route;
 import java.util.List;
 import java.util.Objects;
 
+
 public class RouteOnRoadNet {
     private final List<RouteSegment> segments;
+    public final double length;
+    public final RouteStatus status;
 
-    public RouteOnRoadNet(List<RouteSegment> segments) {
+    public RouteOnRoadNet(List<RouteSegment> segments, double length, RouteStatus status) {
         this.segments = segments;
+        this.length = length;
+        this.status = status;
     }
 
     public List<RouteSegment> getSegments() {
         return segments;
+    }
+
+    public double getLength() {
+        return length;
+    }
+
+    public RouteStatus getStatus() {
+        return status;
     }
 
     @Override